+++ wasatch/FirmwareRequirements.py
@@ -21,6 +21,7 @@ class FirmwareRequirements:
             "get_ble_firmware_version":         { "microcontroller": { "min": "1.0.4.5", "unsupported": [ "11.3.0.37", "1.0.33.7" ] } },
             "get_laser_warning_delay_sec":      { "microcontroller": { "min": "1.0.4.5", "unsupported": [ "11.3.0.37" ] } },
             "hamamatsu_vertical_roi":           { "microcontroller": { "min": "10.0.0.47" } }, # , "fpga": { "min": "35_12_0", "includes": "_" } },
+            "xs_area_scan_offset_kludge":       { "fpga": { "min": "01_04_01", "max": "01_04_30", "includes": "_" } },
         }
 
     def supports(self, feature):
@@ -35,29 +36,38 @@ class FirmwareRequirements:
         fpga_ver  = self.settings.fpga_firmware_version
 
         reqts = self.feature_versions[feature]
+
         if "microcontroller" in reqts:
             reqt = reqts["microcontroller"]
-            if "min" in reqt:
-                min_ = reqt["min"]
-                if vercmp(micro_ver, min_) < 0:
-                    # log.debug(f"supports: {feature} NOT supported (micro {micro_ver} < required {min_}")
-                    return False
             if "unsupported" in reqt:
                 if micro_ver in reqt["unsupported"]:
                     return False
-            # could support "max", list etc
+            if "min" in reqt:
+                min_ = reqt["min"]
+                if vercmp(micro_ver, min_) < 0:
+                    # log.debug(f"supports: {feature} NOT supported (micro {micro_ver} < min {min_}")
+                    return False
+            if "max" in reqt:
+                max_ = reqt["max"]
+                if vercmp(micro_ver, max_) > 0:
+                    # log.debug(f"supports: {feature} NOT supported (micro {micro_ver} > max {max_}")
+                    return False
 
         if "fpga" in reqts:
             reqt = reqts["fpga"]
             if "includes" in reqt:
-                if reqs["includes"] not in fpga_ver:
+                if reqt["includes"] not in fpga_ver:
                     return False
             if "min" in reqt:
                 min_ = reqt["min"]
                 if vercmp(fpga_ver, min_) < 0:
-                    # log.debug(f"supports: {feature} NOT supported (fpga {fpga_ver} < required {min_}")
+                    # log.debug(f"supports: {feature} NOT supported (fpga {fpga_ver} < min {min_}")
                     return False
-            # could support "max", list etc
+            if "max" in reqt:
+                max_ = reqt["max"]
+                if vercmp(fpga_ver, max_) > 0:
+                    # log.debug(f"supports: {feature} NOT supported (fpga {fpga_ver} > max {max_}")
+                    return False
 
         # log.debug(f"supports: {feature} supported")
         return True