+++ wasatch/__init__.py
@@ -1,4 +1,4 @@
-"""This package is a driver for Wasatch Photonics spectrometers"""
+"""This package provides control of Wasatch Photonics spectrometers"""
 
-__version__ = "2.3.0"  # This is used by flit and other pypi things
-version = __version__   # This is to avoid breaking other files that originally used .version instead of __version__
+__version__ = "2.3.2"   # used by flit and other pypi things
+version = __version__   # legacy alias